+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/audiodevice/MicrophoneAudioDeviceList.java
@@ -6,6 +6,9 @@ import de.maxhenkel.voicechat.voice.client.microphone.MicrophoneManager;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TranslationTextComponent;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.function.Supplier;
 
 public class MicrophoneAudioDeviceList extends AudioDeviceList {
 
@@ -18,6 +21,11 @@ public class MicrophoneAudioDeviceList extends AudioDeviceList {
         icon = MICROPHONE_ICON;
         configEntry = VoicechatClient.CLIENT_CONFIG.microphone;
         setAudioDevices(MicrophoneManager.deviceNames());
+    }
+
+    @Override
+    public AudioDeviceEntry createAudioDeviceEntry(String device, Component name, @Nullable ResourceLocation icon, Supplier<Boolean> isSelected) {
+        return new AudioDeviceEntry(device, name, icon, isSelected);
     }
 
 }