+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/audiodevice/SpeakerAudioDeviceList.java
@@ -6,6 +6,9 @@ import de.maxhenkel.voicechat.voice.client.SoundManager;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TranslationTextComponent;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.function.Supplier;
 
 public class SpeakerAudioDeviceList extends AudioDeviceList {
 
@@ -18,6 +21,11 @@ public class SpeakerAudioDeviceList extends AudioDeviceList {
         icon = SPEAKER_ICON;
         configEntry = VoicechatClient.CLIENT_CONFIG.speaker;
         setAudioDevices(SoundManager.getAllSpeakers());
+    }
+
+    @Override
+    public AudioDeviceEntry createAudioDeviceEntry(String device, Component name, @Nullable ResourceLocation icon, Supplier<Boolean> isSelected) {
+        return new SpeakerAudioDeviceEntry(device, name, icon, isSelected);
     }
 
 }