+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/widgets/MicTestButton.java
@@ -133,7 +133,7 @@ public class MicTestButton extends ToggleImageButton implements ImageButton.Tool
 
             SoundManager soundManager;
             if (client == null) {
-                soundManager = new SoundManager(VoicechatClient.CLIENT_CONFIG.speaker.get());
+                soundManager = new SoundManager();
                 ownSoundManager = soundManager;
             } else {
                 soundManager = client.getSoundManager();