+++ fabric/changelog.md
@@ -28,3 +28,4 @@
 - Reworked onboarding
 - Reworked settings screen
 - Increased maximum voice chat volume to 300%
+- Added test speaker button in speaker selection list